requirements.txt
fastmcp>=0.2.0
databricks-sdk>=0.18.0
pydantic>=2.0.0
httpx>=0.25.0
openai>=1.0.0
uvicorn>=0.24.0
